requirements.txt
aiofiles==25.1.0
aioredis==2.0.1
aiosmtplib==3.0.2
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
asgiref==3.9.2
async-timeout==5.0.1
asyncpg==0.30.0
b2sdk==2.10.0
bcrypt==4.3.0
beanie==2.0.0
billiard==4.2.2
blinker==1.9.0
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dnspython==1.16.0
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.10
fastapi-cloud-cli==0.1.5
fastapi-mail==1.5.0
flower==2.0.1
fpdf==1.7.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
humanize==4.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jose==1.0.0
kombu==5.5.4
lazy-model==0.3.0
libpcap==1.11.0b25
logfury==1.0.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.1
numpy==2.3.4
packaging==25.0
passlib==1.7.4
pkg_about==2.0.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psycopg2==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
scapy==2.6.1
sentry-sdk==2.35.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.24
starlette==0.47.3
stripe==13.0.1
tornado==6.5.2
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.14
